pyproject.toml
[project]
name = "castillo-telegram-bot"
version = "0.1.0"
description = "Telegram OCR bot para registrar tickets del restaurante en Google Sheets"
requires-python = ">=3.11"
dependencies = [
    "modal",
    "fastapi",
    "httpx",
    "pydantic>=2",
    "google-api-python-client",
    "google-auth",
]